Move FileSystemLayout to /container module
