+++ container/src/main/java/org/wildfly/swarm/internal/FileSystemLayout.java
@@ -1,4 +1,4 @@
-package org.wildfly.swarm.bootstrap.util;
+package org.wildfly.swarm.internal;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
