+++ resource-adapters/src/main/java/org/wildfly/swarm/resource/adapters/internal/DefaultRarDeploymentFactory.java
@@ -25,7 +25,7 @@ import java.nio.file.attribute.BasicFileAttributes;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.FileAsset;
-import org.wildfly.swarm.bootstrap.util.FileSystemLayout;
+import org.wildfly.swarm.internal.FileSystemLayout;
 import org.wildfly.swarm.resource.adapters.RARArchive;
 import org.wildfly.swarm.spi.api.DefaultDeploymentFactory;
 import org.wildfly.swarm.spi.api.DependenciesContainer;
